Move Organization, Membership and Sponsorship views, forms, filters
New location: fiscal app.
Tests were moved as well, and URLs updated.
