+++ amy/workshops/views.py
@@ -31,7 +31,6 @@ from django.db.models import (
     Sum,
     Prefetch,
 )
-from django.db.models.functions import Now
 from django.forms import HiddenInput
 from django.http import Http404, HttpResponse, JsonResponse
 from django.http import HttpResponseBadRequest
@@ -51,6 +50,7 @@ from api.filters import (
     LearnersOverTimeFilter,
 )
 from api.views import ReportsViewSet
+from fiscal.forms import SponsorshipForm
 from workshops.base_views import (
     AMYCreateView,
     AMYUpdateView,
@@ -63,9 +63,6 @@ from workshops.base_views import (
 )
 from workshops.filters import (
     EventFilter,
-    OrganizationFilter,
-    MembershipFilter,
-    MembershipTrainingsFilter,
     PersonFilter,
     TaskFilter,
     AirportFilter,
@@ -83,15 +80,12 @@ from workshops.forms import (
     AwardForm,
     PersonPermissionsForm,
     PersonsSelectionForm,
-    OrganizationForm,
     BootstrapHelper,
     AdminLookupForm,
-    MembershipForm,
     EventsSelectionForm,
     EventsMergeForm,
     PersonsMergeForm,
     PersonCreateForm,
-    SponsorshipForm,
     TrainingProgressForm,
     BulkAddTrainingProgressForm,
     BulkDiscardProgressesForm,
@@ -154,134 +148,6 @@ def changes_log(request):
         'log': log
     }
     return render(request, 'workshops/changes_log.html', context)
-
-
-# ------------------------------------------------------------
-
-
-class AllOrganizations(OnlyForAdminsMixin, AMYListView):
-    context_object_name = 'all_organizations'
-    template_name = 'workshops/all_organizations.html'
-    filter_class = OrganizationFilter
-    queryset = Organization.objects.prefetch_related(Prefetch(
-        'membership_set',
-        to_attr='current_memberships',
-        queryset=Membership.objects.filter(
-            agreement_start__lte=Now(),
-            agreement_end__gte=Now(),
-        )
-    ))
-    title = 'All Organizations'
-
-
-class OrganizationDetails(OnlyForAdminsMixin, AMYDetailView):
-    queryset = Organization.objects.all()
-    context_object_name = 'organization'
-    template_name = 'workshops/organization.html'
-    slug_field = 'domain'
-    slug_url_kwarg = 'org_domain'
-
-    def get_context_data(self, **kwargs):
-        context = super().get_context_data(**kwargs)
-        context['title'] = 'Organization {0}'.format(self.object)
-        return context
-
-
-class OrganizationCreate(OnlyForAdminsMixin, PermissionRequiredMixin,
-                         AMYCreateView):
-    permission_required = 'workshops.add_organization'
-    model = Organization
-    form_class = OrganizationForm
-
-
-class OrganizationUpdate(OnlyForAdminsMixin, PermissionRequiredMixin,
-                         AMYUpdateView):
-    permission_required = 'workshops.change_organization'
-    model = Organization
-    form_class = OrganizationForm
-    slug_field = 'domain'
-    slug_url_kwarg = 'org_domain'
-
-
-class OrganizationDelete(OnlyForAdminsMixin, PermissionRequiredMixin,
-                         AMYDeleteView):
-    model = Organization
-    slug_field = 'domain'
-    slug_url_kwarg = 'org_domain'
-    permission_required = 'workshops.delete_organization'
-    success_url = reverse_lazy('all_organizations')
-
-
-# ------------------------------------------------------------
-
-
-class AllMemberships(OnlyForAdminsMixin, AMYListView):
-    context_object_name = 'all_memberships'
-    template_name = 'workshops/all_memberships.html'
-    filter_class = MembershipFilter
-    queryset = Membership.objects.annotate(
-        instructor_training_seats_total=(
-            F('seats_instructor_training') +
-            F('additional_instructor_training_seats')
-        ),
-        # for future reference, in case someone would want to implement
-        # this annotation
-        # instructor_training_seats_utilized=(
-        #     Count('task', filter=Q(task__role__name='learner'))
-        # ),
-        instructor_training_seats_remaining=(
-            F('seats_instructor_training') +
-            F('additional_instructor_training_seats') -
-            Count('task', filter=Q(task__role__name='learner'))
-        ),
-    )
-    title = 'All Memberships'
-
-
-class MembershipDetails(OnlyForAdminsMixin, AMYDetailView):
-    queryset = (
-        Membership.objects
-                  .select_related('organization')
-                  .prefetch_related('task_set')
-    )
-    context_object_name = 'membership'
-    template_name = 'workshops/membership.html'
-    pk_url_kwarg = 'membership_id'
-
-    def get_context_data(self, **kwargs):
-        context = super().get_context_data(**kwargs)
-        context['title'] = '{0}'.format(self.object)
-        return context
-
-
-class MembershipCreate(OnlyForAdminsMixin, PermissionRequiredMixin,
-                       PrepopulationSupportMixin, AMYCreateView):
-    permission_required = [
-        'workshops.add_membership',
-        'workshops.change_organization',
-    ]
-    model = Membership
-    form_class = MembershipForm
-    populate_fields = ['organization']
-
-
-class MembershipUpdate(OnlyForAdminsMixin, PermissionRequiredMixin,
-                       RedirectSupportMixin, AMYUpdateView):
-    permission_required = 'workshops.change_membership'
-    model = Membership
-    form_class = MembershipForm
-    pk_url_kwarg = 'membership_id'
-
-
-class MembershipDelete(OnlyForAdminsMixin, PermissionRequiredMixin,
-                       AMYDeleteView):
-    model = Membership
-    permission_required = 'workshops.delete_membership'
-    pk_url_kwarg = 'membership_id'
-
-    def get_success_url(self):
-        return reverse('organization_details', args=[
-            self.get_object().organization.domain])
 
 
 # ------------------------------------------------------------
@@ -1330,27 +1196,6 @@ def event_dismiss_metadata_changes(request, slug):
                      'Changes to {} were dismissed.'.format(event.slug))
 
     return redirect(reverse('event_details', args=[event.slug]))
-
-
-class SponsorshipCreate(OnlyForAdminsMixin, PermissionRequiredMixin,
-                        AMYCreateView):
-    model = Sponsorship
-    permission_required = 'workshops.add_sponsorship'
-    form_class = SponsorshipForm
-
-    def get_success_url(self):
-        return reverse('event_edit', args=[self.object.event.slug]) + \
-            '#sponsors'
-
-
-class SponsorshipDelete(OnlyForAdminsMixin, PermissionRequiredMixin,
-                        AMYDeleteView):
-    model = Sponsorship
-    permission_required = 'workshops.delete_sponsorship'
-
-    def get_success_url(self):
-        return reverse('event_edit', args=[self.get_object().event.slug]) + \
-            '#sponsors'
 
 # ------------------------------------------------------------
